+++ src/octree/raytracing/types.rs
@@ -15,7 +15,7 @@ pub(crate) struct NodeStackItem {
     pub(crate) bounds_intersection: CubeRayIntersection,
     pub(crate) bounds: Cube,
     pub(crate) node: u32,
-    pub(crate) target_octant: u32,
+    pub(crate) target_octant: u8,
     pub(crate) child_center: V3c<f32>,
 }
